these tests are only for windows
